BGRemover/views.py
from django.shortcuts import render
from django.contrib.auth.decorators import login_required
from django.views.decorators.csrf import csrf_exempt
from django.http import HttpResponse


import cv2
import numpy as np

from PIL import Image
from io import BytesIO




BLUR = 21
CANNY_THRESH_1 = 5
CANNY_THRESH_2 = 400
MASK_DILATE_ITER = 10
MASK_ERODE_ITER = 10
MASK_COLOR = (1.0,1.0,1.0) # In BGR format

# Create your views here.
@login_required(login_url='register')
def home(request):
    return render(request, 'BGRemover/index.html', locals())

@csrf_exempt
def removeBgProcess(request):
    img_str = request.FILES['image'].read()

    nparr = np.fromstring(img_str, np.uint8)
    img = cv2.imdecode(nparr, cv2.IMREAD_COLOR)

    gray = cv2.cvtColor(img,cv2.COLOR_BGR2GRAY)

    edges = cv2.Canny(gray, CANNY_THRESH_1, CANNY_THRESH_2)
    edges = cv2.dilate(edges, None)
    edges = cv2.erode(edges, None)

    contour_info = []
    contours, _ = cv2.findContours(edges, cv2.RETR_LIST, cv2.CHAIN_APPROX_NONE)

    for c in contours:
        contour_info.append((
            c,
            cv2.isContourConvex(c),
            cv2.contourArea(c),
        ))
    contour_info = sorted(contour_info, key=lambda c: c[2], reverse=True)
    max_contour = contour_info[0]

    mask = np.zeros(edges.shape)
    cv2.fillConvexPoly(mask, max_contour[0], (255))

    mask = cv2.dilate(mask, None, iterations=MASK_DILATE_ITER)
    mask = cv2.erode(mask, None, iterations=MASK_ERODE_ITER)
    mask = cv2.GaussianBlur(mask, (BLUR, BLUR), 0)
    mask_stack = np.dstack([mask]*3)  

    mask_stack  = mask_stack.astype('float32') / 255.0      
    img         = img.astype('float32') / 255.0                 

    masked = (mask_stack * img) + ((1-mask_stack) * MASK_COLOR)
    masked = (masked * 255).astype('uint8')

    masked = masked[...,::-1].copy()

    maskImage = Image.fromarray(masked, 'RGB')
    response = HttpResponse(content_type="image/jpeg")
    maskImage.save(response, "JPEG")
    return response