requirements.txt
gunicorn
django-heroku
Django~=2.0
psycopg2
gTTS
numpy
opencv-python
Pillow